+++ vector/src/main/java/im/vector/app/features/MainActivity.kt
@@ -31,7 +31,7 @@ import im.vector.app.core.error.ErrorFormatter
 import im.vector.app.core.extensions.startSyncing
 import im.vector.app.core.platform.VectorBaseActivity
 import im.vector.app.core.utils.deleteAllFiles
-import im.vector.app.databinding.FragmentLoadingBinding
+import im.vector.app.databinding.ActivityMainBinding
 import im.vector.app.features.home.HomeActivity
 import im.vector.app.features.home.ShortcutsHandler
 import im.vector.app.features.login.LoginActivity
@@ -66,7 +66,7 @@ data class MainActivityArgs(
  * This Activity, when started with argument, is also doing some cleanup when user signs out,
  * clears cache, is logged out, or is soft logged out
  */
-class MainActivity : VectorBaseActivity<FragmentLoadingBinding>(), UnlockedActivity {
+class MainActivity : VectorBaseActivity<ActivityMainBinding>(), UnlockedActivity {
 
     companion object {
         private const val EXTRA_ARGS = "EXTRA_ARGS"
@@ -81,7 +81,7 @@ class MainActivity : VectorBaseActivity<FragmentLoadingBinding>(), UnlockedActiv
         }
     }
 
-    override fun getBinding() = FragmentLoadingBinding.inflate(layoutInflater)
+    override fun getBinding() = ActivityMainBinding.inflate(layoutInflater)
 
     private lateinit var args: MainActivityArgs